requirements.txt
requests~=2.27.1
web3~=5.29.0
discord~=1.7.3

pandas~=1.4.2
python-dotenv~=0.20.0